third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_hit_test_scroll_visible_descendant-expected.txt

This is a testharness.js-based test.
FAIL Wheel-scroll over pointer-events: auto descendant scrolls pointer-events: none scroller. assert_equals: Incorrect element scrolled expected "overlay" but got "root"
Harness: the test ran to completion.
